+++ src/components/antd/main.js
@@ -6,4 +6,3 @@ export * from './Table';
 export * from './Timeline';
 export * from './MenuItem';
 export * from './Collapse';
-export * from './SimpleProgressBar';