+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.2
 
 - Add tests for and update `updatedQuery` function [#3664093](https://github.com/SpotlightData/nanowire-extensions/pull/78/commits/3664093f5315827a4b7b4e869c3aa9ece0100f0b)
+- Set fixed uppy version [#9f0fc33](https://github.com/SpotlightData/nanowire-extensions/pull/78/commits/9f0fc335475e4d68bc3be043fb1511bb8082ebe6)
 
 # 1.3.0